retrieve_all and retrieve_one controllers up and running seamlessly.
had to do a mod on the doc provided by sequelize.  they dont support raw queries (in terms of searching for a specific medium in retrieve_one CB) had to send in oobject to search.  Although, I have a feeling that the reason it didnt work correctly was because I was passing in a variable not a hard coded ID to search for.  Ima play with it for a sec.

Anyway, this commit is running a-okay.

UPDATE: nope, raw queries are def deprecated. variable or no variable.  Must do what I did in terms of using an object to search for.